脚本/新闻蓝海报机器人/requirements.txt
pyTelegramBotAPI
Pillow
python-dotenv
google-genai
